backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
python-multipart>=0.0.9,<1.0
pydantic>=2.8,<3.0
pydantic-settings>=2.3,<3.0
python-dotenv>=1.0,<2.0
chromadb>=0.5,<1.0
litellm>=1.51,<2.0
pypdf>=5.0,<6.0
pytesseract>=0.3,<1.0
Pillow>=10.4,<11.0
faster-whisper>=1.0,<2.0
httpx>=0.27,<1.0
aiosqlite>=0.20,<1.0
